requirements.txt
# Core DL
torch>=2.1
torchvision>=0.16
transformers>=4.44

# Data / preprocessing
numpy>=1.24
Pillow>=10.0
dpkt>=1.9.8
PyYAML>=6.0
tqdm>=4.66

# Evaluation / visualization
scikit-learn>=1.3
matplotlib>=3.8

# Optional ensemble meta learners
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2
